requirements.txt
pyyaml
requests
socket
